API issues

+++ src/main/java/com/iemr/helpline1097/repository/co/beneficiary/BenCalServiceCatSubcatMappingRepo.java
@@ -24,11 +24,12 @@ package com.iemr.helpline1097.repository.co.beneficiary;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RestResource;
 import org.springframework.stereotype.Repository;
+import java.math.BigInteger;
 
 import com.iemr.helpline1097.data.co.beneficiarycall.BenCallServicesMappingHistory;
 
 @Repository
 @RestResource(exported = false)
-public interface BenCalServiceCatSubcatMappingRepo extends CrudRepository<BenCallServicesMappingHistory, Long> {
+public interface BenCalServiceCatSubcatMappingRepo extends CrudRepository<BenCallServicesMappingHistory, BigInteger> {
 
 }
